enums i egna filer och all swagger-info i property-filen
