+++ current-race/pom.xml
@@ -22,6 +22,7 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <java.version>1.8</java.version>
         <artifactFinalName>cag-${project.artifactId}</artifactFinalName>
+        <tomcat.version>8.0.28</tomcat.version>
     </properties>
 
     <dependencies>
