+++ current-race/src/main/java/se/cag/labs/currentrace/apicontroller/SystemController.java
@@ -2,7 +2,7 @@ package se.cag.labs.currentrace.apicontroller;
 
 import org.springframework.web.bind.annotation.*;
 import se.cag.labs.currentrace.apicontroller.apimodel.VersionResponse;
-import se.cag.labs.currentrace.util.BuildInfo;
+import se.cag.labs.currentrace.config.BuildInfo;
 
 @RestController
 @CrossOrigin(origins = "*", methods = {RequestMethod.GET})
