+++ current-race/src/main/java/se/cag/labs/currentrace/config/SwaggerConfig.java
@@ -2,7 +2,6 @@ package se.cag.labs.currentrace.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import se.cag.labs.currentrace.util.BuildInfo;
 import springfox.documentation.builders.RequestHandlerSelectors;
 import springfox.documentation.service.ApiInfo;
 import springfox.documentation.spi.DocumentationType;
@@ -27,7 +26,8 @@ public class SwaggerConfig {
 
     private ApiInfo apiInfo() {
         return new ApiInfo(
-                "Current Race! API", "API for Current Race",
+                BuildInfo.getTitle(),
+                BuildInfo.getDescription(),
                 BuildInfo.getVersion(),
                 "",
                 "",
