+++ current-race/src/main/java/se/cag/labs/currentrace/services/PassageDetectedService.java
@@ -4,48 +4,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import se.cag.labs.currentrace.services.repository.CurrentRaceRepository;
 import se.cag.labs.currentrace.services.repository.datamodel.RaceStatus;
-
-import java.util.HashMap;
-import java.util.Map;
+import se.cag.labs.currentrace.services.statuses.PassageDetectedStatus;
+import se.cag.labs.currentrace.services.statuses.SensorType;
 
 
 @Service
 public class PassageDetectedService {
     @Autowired
     private CurrentRaceRepository repository;
-
-    public enum SensorType {
-        START("START_ID"),
-        MIDDLE("MIDDLE_ID"),
-        FINISH("FINISH_ID");
-
-        private final String id;
-        private static final Map<String, SensorType> lookup = new HashMap<>();
-
-        static {
-            for (SensorType type : SensorType.values()) {
-                lookup.put(type.getId(), type);
-            }
-        }
-
-        SensorType(String id) {
-            this.id = id;
-        }
-
-        public String getId() {
-            return id;
-        }
-
-        public static SensorType get(String id) {
-            return lookup.get(id);
-        }
-    }
-
-    public enum PassageDetectedStatus {
-        ACCEPTED,
-        IGNORED,
-        ERROR
-    }
 
     public PassageDetectedStatus passageDetected(String sensorID, long timestamp) {
         RaceStatus raceStatus = repository.findByRaceId(RaceStatus.ID);
