+++ current-race/src/main/java/se/cag/labs/currentrace/services/statuses/CancelRaceServiceReturnStatus.java
@@ -1,0 +1,6 @@
+package se.cag.labs.currentrace.services.statuses;
+
+public enum CancelRaceServiceReturnStatus {
+    ACCEPTED,
+    NOT_FOUND
+}
