+++ current-race/src/main/java/se/cag/labs/currentrace/services/statuses/PassageDetectedStatus.java
@@ -1,0 +1,7 @@
+package se.cag.labs.currentrace.services.statuses;
+
+public enum PassageDetectedStatus {
+    ACCEPTED,
+    IGNORED,
+    ERROR
+}
