+++ current-race/src/main/java/se/cag/labs/currentrace/services/statuses/StartRaceReturnStatus.java
@@ -1,0 +1,6 @@
+package se.cag.labs.currentrace.services.statuses;
+
+public enum StartRaceReturnStatus {
+    STARTED,
+    FOUND
+}
