+++ current-race/src/main/resources/BuildInfo.properties
@@ -1,2 +1,4 @@
 # Deployment scripts depend on the name pom.version - do not change it to project.version.
 AppVersion=@project.version@
+title=Current Race! API
+description=API for Current Race
